completed crud user functionalities and protected all goal routes + get user route

+++ backend/controlllers/goalController.js
@@ -1,13 +1,15 @@
 const asyncHandler = require("express-async-handler")
 const Goal=require("../models/goalsModel")
+const User = require("../models/userModel");
+
 
 const getGoals=asyncHandler(async(req,res)=>{
-    if(!req.body.text){
-res.status(400)
-
-throw new Error(`Please add a text field`)
-
-    }    
+   
+    const user=await User.findById(req.user.id)
+    if(!user){
+        res.status(401)
+        throw new Error(`User not found`)
+    }
     const goals=await Goal.find()
     res.status(200).json({
         message:`goals fetched`,
@@ -18,16 +20,32 @@ throw new Error(`Please add a text field`)
 
 
 const setGoals=asyncHandler(async(req,res)=>{
+    const user= await User.findById(req.user.id)
+     if(!user){
+        res.status(401)
+        throw new Error(`User not found`);
+
+     }
+
     const goals=req.body.text
+
     const createGoals=await Goal.create({
       text:goals
     })
+
     res.status(200).json({message:`Post some freaking goals`,
      goals
 })
 })
 
 const addGoals=asyncHandler(async(req,res)=>{
+    const user= await User.findById(req.user.id)
+     if(!user){
+        res.status(401)
+        throw new Error(`User not found`);
+
+     }
+
     const createdgoals=req.params.id
     const findGoal=await Goal.findById(createdgoals)
 
@@ -45,11 +63,21 @@ const addGoals=asyncHandler(async(req,res)=>{
 
 
 const deleteGoals = asyncHandler(async (req, res) => {
+    const user= await User.findById(req.user.id)
+
+    if(!user){
+       res.status(401)
+       throw new Error(`User not found`);
+
+    }
+    // console.log('req.params.id:', req.params.id);
+    
     const goal=await Goal.findById(req.params.id)
     if(!goal){
         res.status(400)
         throw new Error('Goal not found')
     }
+
     const result=await Goal.deleteOne({_id:goal})
     res.status(200).json({
        id: result
